+++ elmyra.ip.access.epo/CHANGES.rst
@@ -13,6 +13,7 @@ development
 - ui: major refactoring of javascript code from core.js to components/*
 - ui: indicate activity (spinner) while fetching document details
 - middleware: run document acquisition when document fulltext details yielded no results at DEPATISconnect
+- middleware: fall back to CIPO for Canadian drawings
 
 0.37.0
 ======